moved completed adventure section to components
